refactor to avoid forking
This does a few things:

- Updates to Wizer 2.0.0 so we can use Wasmtime 8.0.1 everywhere and no longer need to put the tests in a separate crate
- Bumps the Wasmtime version for the `preview2` implementation to match
- Uses `Wizer::make_linker` to enable a custom-configured `WasiCtx` so we don't have to fork a process to fine-tune the set of environment variables and ensure stdio descriptors are files (which is needed to make CPython happy on Windows)
- Converts the crate to a library in preparation for making it available in Python as either a library or a command
