+++ Cargo.toml
@@ -7,9 +7,9 @@ edition = "2021"
 anyhow = "1.0.68"
 clap = { version = "4.1.4", features = [ "derive" ] }
 tar = "0.4.38"
-tempfile = "3.3.0"
-wizer = "1.6.0"
-zstd = "0.10.0"
+tempfile = "3.5.0"
+wizer = "2.0.0"
+zstd = "0.11.1"
 componentize-py-shared = { path = "shared" }
 wasmparser = "0.102.0"
 wasm-encoder = "0.25.0"
@@ -18,14 +18,34 @@ wit-component = "0.7.4"
 indexmap = "1.9.3"
 bincode = "1.3.3"
 heck = "0.4.1"
+once_cell = "1.17.1"
+wasmtime-wasi = "8.0.1"
+wasi-common = "8.0.1"
+wasmtime = "8.0.1"
+cap-std = "1.0.15"
+
+[dev-dependencies]
+tokio = { version = "1.27.0", features = ["macros", "rt", "rt-multi-thread", "fs"] }
+wasi-host = { package = "host", path = "preview2/host" }
+wasmtime-wasi-preview2 = { package = "wasi-cap-std-sync", path = "preview2/wasi-common/cap-std-sync" }
+wasi-preview2 = { package = "wasi-common", path = "preview2/wasi-common" }
+async-trait = "0.1.68"
+once_cell = "1.17.1"
+proptest = "1.1.0"
+hex = "0.4.3"
 
 [build-dependencies]
 anyhow = "1.0.68"
 tar = "0.4.38"
-zstd = "0.10.0"
+zstd = "0.11.1"
+test-generator = { path = "test-generator" }
 
 [workspace]
 members = [
   "runtime",
   "shared",
+  "test-generator",
+]
+exclude = [
+  "preview2"
 ]
