+++ src/test/echoes_generated.rs
@@ -1,1 +1,2 @@
+#![allow(warnings)]
 include!(concat!(env!("OUT_DIR"), "/echoes_generated.rs"));
