+++ test-generator/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "test-generator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.70"
+getrandom = "0.2.9"
+hex = "0.4.3"
+proptest = "1.1.0"
