--- test/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "componentize-py-test"
-version = "0.1.0"
-edition = "2021"
-
-[dev-dependencies]
-anyhow = { version = "1.0.70", features = ["backtrace"] }
-tempfile = "3.5.0"
-wasmtime = { version = "7.0.0", features = ["component-model"] }
-tokio = { version = "1.27.0", features = ["macros", "rt", "rt-multi-thread", "fs"] }
-wasi-host = { package = "host", path = "../preview2/host" }
-wasmtime-wasi-preview2 = { package = "wasi-cap-std-sync", path = "../preview2/wasi-common/cap-std-sync" }
-wasi-preview2 = { package = "wasi-common", path = "../preview2/wasi-common" }
-async-trait = "0.1.68"
-once_cell = "1.17.1"
-proptest = "1.1.0"
-hex = "0.4.3"
-
-[build-dependencies]
-anyhow = "1.0.70"
-getrandom = "0.2.9"
-hex = "0.4.3"
-proptest = "1.1.0"
-
-[workspace]
